requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Gemini API依赖
google-generativeai>=0.8.0

langsmith==0.1.147
langchain-core==0.1.53
langchain==0.2.0
langchain-community==0.2.0

# 实用工具依赖
beautifulsoup4>=4.12.0
duckduckgo-search>=3.9.0
qrcode>=7.4.0
pillow>=10.0.0
psutil>=5.9.0
pandas>=2.0.0
lxml>=4.9.0
